// api-client/Api.ts
export interface TodoResponseDto {
  id: string;
  content: string;
  dueDate?: string;
  createdAt: string;
}